bump axios 0.21.4 to 0.24.0
this was quite some story. In Axios 0.22.0 they broke typescript
completely. Then in 0.23.0 and 0.24.0 they changed between the type
never and any and then also a type for the headers was introduced.

Also once more the chromedriver needed to be updated to the latest
version. I use now an action to set it up for Github Actions.